Scripts/BackEnd/requirements.txt
Flask==2.3.3
python-pptx==0.6.22
